Using the new grid implemented for the new homepage
Gap-x decreased by 2 (6 to 4) (is this to small)
All col-start/end needed to be increased by 1
