+++ src/components/layout/Layout.tsx
@@ -9,7 +9,7 @@ interface Props {
 export const Layout = ({ rowsSetup, children, dataQa }: Props) => (
   <div
     data-qa={dataQa}
-    className={`container grid ${rowsSetup} grid-cols-24 gap-6`}
+    className={`grid ${rowsSetup} grid-cols-container gap-y-6 gap-x-2 lg:gap-x-4`}
   >
     {children}
   </div>
