+++ src/components/noResults/NoSearchResults.tsx
@@ -17,7 +17,7 @@ const NoResultsCopy = ({ header, copy }) => {
 
 export const NoSearchResults = ({ areFiltersApplied, query }: Props) => {
   return (
-    <div className="flex flex-col items-center col-start-7 col-end-26 text-lg text-gray-800 mt-10 ">
+    <div className="flex flex-col items-center col-start-8 col-end-26 text-lg text-gray-800 mt-10 ">
       <NoResultsIcon className="h-80 w-80 mb-6" />
       {areFiltersApplied ? (
         <NoResultsCopy
