+++ src/components/searchHero/SearchHero.tsx
@@ -29,14 +29,14 @@ const SearchHero = () => {
 
   return (
     <>
-      <div className="col-start-1 col-end-25 row-start-2 row-end-2 bg-primary-light rounded-lg" />
-      <div className="col-start-3 col-end-15 md:col-start-4 md:col-end-15 lg:col-start-4 lg:col-end-15 row-start-2 row-end-2 self-center">
+      <div className="col-start-2 col-end-26 row-start-2 row-end-2 bg-primary-light rounded-lg" />
+      <div className="col-start-4 col-end-16 md:col-start-5 md:col-end-16 lg:col-start-5 lg:col-end-16 row-start-2 row-end-2 self-center">
         <h1 className="mb-8 text-4xl font-medium">
           What videos do you need today?
         </h1>
         <Search size="big" showIconOnly={false} onSearch={onSearch} />
       </div>
-      <div className="col-start-16 col-end-25 row-start-2 row-end-2 self-center justify-self-start z-0">
+      <div className="col-start-17 col-end-26 row-start-2 row-end-2 self-center justify-self-start z-0">
         <HomeImageSVG />
       </div>
     </>
