+++ src/components/searchResults/SearchResults.tsx
@@ -39,7 +39,7 @@ export function SearchResults({
   };
 
   return (
-    <div className="col-start-6 col-end-26">
+    <div className="col-start-8 col-end-26">
       {query && <Helmet title={`${query} videos`} />}
       <SearchResultsSummary
         count={results?.pageSpec?.totalElements}
