+++ src/views/homepageWithCategories/ResponsiveHomeView.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import DisciplinesWidget from 'src/components/disciplinesWidget/DisciplinesWidget';
-import ResponsiveLayout from 'src/components/layout/ResponsiveLayout';
-import NavbarResponsive from 'src/components/layout/NavbarResponsive';
+import Navbar from 'src/components/layout/Navbar';
 import Footer from 'src/components/layout/Footer';
+import { Layout } from 'src/components/layout/Layout';
 
 const ResponsiveHomeView = () => {
   return (
-    <ResponsiveLayout rowsSetup="grid-rows-home-responsive">
-      <NavbarResponsive />
+    <Layout rowsSetup="grid-rows-home-responsive">
+      <Navbar />
       <DisciplinesWidget />
       <Footer columnPosition="col-start-2 col-end-26" />
-    </ResponsiveLayout>
+    </Layout>
   );
 };
